Update: code reorganization
