+++ src/jvmTest/kotlin/io/smallibs/pilin/standard/either/MonadTest.kt
@@ -1,0 +1,49 @@
+package io.smallibs.pilin.standard.either
+
+import io.smallibs.pilin.laws.Monad.`(a bind f) bind g = a bind {x in f x bind g}`
+import io.smallibs.pilin.laws.Monad.`a bind returns = a`
+import io.smallibs.pilin.laws.Monad.`returns a bind h = h a`
+import io.smallibs.pilin.standard.support.Functions.ret
+import io.smallibs.pilin.standard.support.Functions.retInt
+import io.smallibs.pilin.standard.support.Functions.retStr
+import io.smallibs.pilin.standard.support.constant
+import io.smallibs.pilin.standard.support.either
+import io.smallibs.pilin.type.App
+import io.smallibs.pilin.type.Fun
+import kotlinx.coroutines.runBlocking
+import org.junit.Test
+import org.quicktheories.WithQuickTheories
+
+internal class MonadTest : WithQuickTheories {
+
+    @Test
+    fun `(Either) returns a bind h = h a`() {
+        qt().forAll(integers().all(), either<Unit, String>(Unit)).check { a, r ->
+            runBlocking {
+                Either.monad<Unit>().`returns a bind h = h a`(retStr(r), a)
+            }
+        }
+    }
+
+    @Test
+    fun `(Either) a bind returns = a`() {
+        qt().forAll(either<Unit, Int>(constant(Unit))(integers().all())).check { a ->
+            runBlocking {
+                Either.monad<Unit>().`a bind returns = a`(a)
+            }
+        }
+    }
+
+    @Test
+    fun `(Either) (a bind f) bind g = a bind {x in f x bind g}`() {
+        qt().forAll(either<Unit, Int>(constant(Unit))(integers().all()),
+            either<Unit, String>(Unit),
+            either<Unit, Int>(Unit))
+            .check { a, rf, rg ->
+                runBlocking {
+                    Either.monad<Unit>().`(a bind f) bind g = a bind {x in f x bind g}`(retStr(rf), retInt(rg), a)
+                }
+            }
+    }
+
+}
