+++ src/jvmTest/kotlin/io/smallibs/pilin/standard/either/SelectiveTest.kt
@@ -1,0 +1,21 @@
+package io.smallibs.pilin.standard.either
+
+import io.smallibs.pilin.laws.Selective.`x select (pure id) = fold(id)(id) map x`
+import io.smallibs.pilin.standard.either.Either
+import io.smallibs.pilin.standard.identity.Identity
+import io.smallibs.pilin.standard.option.Option
+import io.smallibs.pilin.standard.support.either
+import kotlinx.coroutines.runBlocking
+import org.junit.Test
+import org.quicktheories.WithQuickTheories
+
+internal class SelectiveTest : WithQuickTheories {
+
+    @Test
+    fun `(Either) x select (pure id) = fold(id)(id) map x`() {
+        qt().forAll(either<Int, Int>(integers().allPositive())(integers().allPositive())).check { a ->
+            runBlocking { Either.selective<Unit>().`x select (pure id) = fold(id)(id) map x`(a) }
+        }
+    }
+
+}
