+++ src/jvmTest/kotlin/io/smallibs/pilin/standard/identity/FunctorTest.kt
@@ -1,4 +1,4 @@
-package io.smallibs.pilin.standard
+package io.smallibs.pilin.standard.identity
 
 import io.smallibs.pilin.laws.Functor.`map (f compose g) = map f compose map g`
 import io.smallibs.pilin.laws.Functor.`map id = id`
@@ -19,8 +19,6 @@ internal class FunctorTest : WithQuickTheories {
     private val str: Fun<Int, String> = { i -> i.toString() }
     private val int: Fun<String, Int> = { i -> i.toInt() }
 
-    // -----------------------------------------------------------------------------------------------------------------
-
     @Test
     fun `(Identity) map id = id `() {
         qt().forAll(identity(integers().all())).check { a ->
@@ -29,39 +27,10 @@ internal class FunctorTest : WithQuickTheories {
     }
 
     @Test
-    fun `(Option) map id = id `() {
-        qt().forAll(option(integers().all())).check { a ->
-            runBlocking { Option.functor.`map id = id`(a) }
-        }
-    }
-
-    @Test
-    fun `(Either) map id = id `() {
-        qt().forAll(either<Unit, Int>(constant(Unit))(integers().all())).check { a ->
-            runBlocking { Either.functor<Unit>().`map id = id`(a) }
-        }
-    }
-
-    // -----------------------------------------------------------------------------------------------------------------
-
-    @Test
     fun `(Identity) map (incr compose toString) = (map incr) compose (map toString) `() {
         qt().forAll(identity(integers().all())).check { a ->
             runBlocking { Identity.functor.`map (f compose g) = map f compose map g`(int, str, a) }
         }
     }
 
-    @Test
-    fun `(Option) map (incr compose toString) = (map incr) compose (map toString) `() {
-        qt().forAll(option(integers().all())).check { a ->
-            runBlocking { Option.functor.`map (f compose g) = map f compose map g`(int, str, a) }
-        }
-    }
-
-    @Test
-    fun `(Either) map (incr compose toString) = (map incr) compose (map toString) `() {
-        qt().forAll(either<Unit, Int>(constant(Unit))(integers().all())).check { a ->
-            runBlocking { Either.functor<Unit>().`map (f compose g) = map f compose map g`(int, str, a) }
-        }
-    }
 }
